+++ crates/rs-rtmp-push/src/session.rs
@@ -665,84 +665,11 @@ const READ_LOOP_IDLE_MS: u64 = 50;
 /// Detection latency for server-initiated errors is bounded by
 /// `READ_LOOP_HOLD_MS + READ_LOOP_IDLE_MS` (~55 ms), which is far below
 /// xiu's 2 s inactivity timer and well under the consumer-task `WRITE_TIMEOUT`.
-async fn read_loop(io: Arc<Mutex<Box<dyn TNetIO + Send + Sync>>>, poisoned: Arc<AtomicBool>) {
-    let mut unpacketizer = ChunkUnpacketizer::new();
-    loop {
-        // Sleep FIRST every iteration so the read-loop is mutex-busy for
-        // at most HOLD/(HOLD+IDLE) = 5/55 ≈ 9 % of the time, regardless
-        // of whether the previous iteration read data or timed out. RTMP
-        // servers send Window-Ack frames frequently during a healthy
-        // push; without an unconditional sleep the loop stays hot when
-        // reads succeed and the mutex stays held ~50 % of the time —
-        // which throttles `send_tag` writers and observed in the #103
-        // E2E run as ~0.4 x output even after every other pacing fix.
-        tokio::time::sleep(Duration::from_millis(READ_LOOP_IDLE_MS)).await;
-
-        // Don't *await* the mutex — `send_tag` is on the hot path and
-        // must win immediately. If contended, skip this round entirely
-        // (the next iteration's sleep already gives writers headroom).
-        let result = match io.try_lock() {
-            Ok(mut guard) => {
-                tokio::time::timeout(Duration::from_millis(READ_LOOP_HOLD_MS), guard.read()).await
-                // guard (and the mutex) is dropped here regardless of outcome.
-            }
-            Err(_) => continue,
-        };
-
-        let data = match result {
-            Err(_timeout) => continue,
-            Ok(Ok(d)) => d,
-            Ok(Err(_)) => {
-                poisoned.store(true, Ordering::Relaxed);
-                return;
-            }
-        };
-
-        if data.is_empty() {
-            poisoned.store(true, Ordering::Relaxed);
-            return;
-        }
-
-        unpacketizer.extend_data(&data[..]);
-        loop {
-            match unpacketizer.read_chunks() {
-                Ok(UnpackResult::Chunks(chunks)) => {
-                    for chunk in chunks {
-                        if let Ok(Some(msg)) = MessageParser::new(chunk).parse() {
-                            match msg {
-                                RtmpMessageData::SetChunkSize { chunk_size } => {
-                                    unpacketizer.update_max_chunk_size(chunk_size as usize);
-                                }
-                                RtmpMessageData::Amf0Command {
-                                    command_name,
-                                    others,
-                                    ..
-                                } if amf_string(&command_name) == "onStatus" => {
-                                    // Watch for mid-stream onStatus errors.
-                                    let is_error = others.iter().any(|v| {
-                                        let Amf0ValueType::Object(m) = v else {
-                                            return false;
-                                        };
-                                        m.get("level")
-                                            .map(|lv| matches!(lv, Amf0ValueType::UTF8String(s) if s == "error"))
-                                            .unwrap_or(false)
-                                    });
-                                    if is_error {
-                                        poisoned.store(true, Ordering::Relaxed);
-                                        return;
-                                    }
-                                }
-                                _ => {}
-                            }
-                        }
-                    }
-                }
-                Err(_) => break,
-                Ok(_) => {}
-            }
-        }
-    }
-}
+// Read loop is large (RTMP message dispatch + Acknowledgement state).
+// Extracted to a sibling file to keep this file under the 1000-line cap.
+#[path = "session_read_loop.rs"]
+mod read_loop_mod;
+use read_loop_mod::read_loop;
 
 // -------------------------------------------------------------------------
 // AMF value helpers
