+++ leptos-ui/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "leptos-ui"
-version = "0.3.77"
+version = "0.3.78"
 edition = "2024"
 license = "MIT"
