+++ src-tauri/tauri.conf.json
@@ -1,7 +1,7 @@
 {
   "$schema": "https://raw.githubusercontent.com/tauri-apps/tauri/dev/crates/tauri-cli/config.schema.json",
   "productName": "Restreamer",
-  "version": "0.3.77",
+  "version": "0.3.78",
   "identifier": "media.newlevel.restreamer",
   "build": {
     "frontendDist": "../dist",
